feat(chatbot): implement DAOP-style conversation flow with product detection and order state machine

+++ src/Entreprenly.WebServices/Chatbot/Domain/Model/ValueObjects/OrderItem.cs
@@ -4,7 +4,7 @@ public record OrderItem(
     int ProductId,
     string ProductName,
     decimal UnitPrice,
-    int Quantity
+    decimal Quantity
 )
 {
     public decimal Subtotal => UnitPrice * Quantity;
